+++ changelog.md
@@ -3,10 +3,12 @@
 ### New Features
 * Add note about serialport node error (thanks @octopicorn)
 * Closes #142 - Graceful termination
+* Adds `commsDownDetection` option, enabled by default, don't need to use 'ready' function any more.
 
 ### Breaking changes
 * Connect rejects when radio system is down on firmware version 2 see #144
-* Module drops all attached event emitters on disconnect see #144
+* Module drops all attached event emitters on disconnect see #142
+* Connect resolved too early, resulted in resolves of systems that could not be user.
 
 ### Bug Fixes
 * Fixes #144 - v2 firmware succeeds when no board attached